# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Bitaqty Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendors/notifications": {
            "get": {
                "tags": [
                    "VendorNotification"
                ],
                "summary": "Notifications",
                "description": "Vendor Notifications",
                "operationId": "VendorNotifications",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "data": {
                                                        "type": "object",
                                                        "example": []
                                                    },
                                                    "is_read": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vendors/notifications/{id}": {
            "delete": {
                "tags": [
                    "VendorNotification"
                ],
                "summary": "Notifications",
                "description": "Vendor Delete Notifications",
                "operationId": "VendorDeleteNotifications",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "VendorNotification"
                ],
                "summary": "Notifications",
                "description": "Vendor Read Notifications",
                "operationId": "VendorReadNotifications",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}